+++ agents-manage-api/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @inkeep/agents-manage-api
 
+## 0.31.6
+
+### Patch Changes
+
+- @inkeep/agents-core@0.31.6
+
+## 0.31.5
+
+### Patch Changes
+
+- @inkeep/agents-core@0.31.5
+
 ## 0.31.4
 
 ### Patch Changes