+++ packages/agents-core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @inkeep/agents-core
 
+## 0.31.6
+
+## 0.31.5
+
 ## 0.31.4
 
 ### Patch Changes